iphone/requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-allauth==0.49.0
django-bootstrap4==21.2
django-ckeditor==6.4.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-mailjet==0.3.1
django-phonenumber-field==6.1.0
django-session-timeout==0.1.0
django-templated-mail==1.1.1
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djoser==2.1.0
idna==3.3
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
mailjet-rest==1.3.4
MarkupSafe==2.1.1
mysql==0.0.3
mysqlclient==2.1.0
oauthlib==3.2.0
phonenumbers==8.12.48
Pillow==9.0.1
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
pycparser==2.21
PyJWT==2.1.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
soupsieve==2.3.1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
uritemplate==4.1.1
urllib3==1.26.8
wincertstore==0.2
